pyproject.toml
[tool.poetry]
name = "slidestopdf"
version = "1.0.0"
description = "Turn your google slides presentation into a pdf!"
authors = ["TanglyEagle7718"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.29.0"
pillow = "^11.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
